intel_sgx: check if SGX is enabled in feature control
Load the SGX driver if and only if the SGX enable bit is set in the
feature control MSR and the feature control MSR is locked.  Do not
write the MSR if it is unlocked as doing so could unintentionally
prevent other other modules from running, e.g. KVM or MCE.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>